exporting helper function to utils
