+++ frontend/src/util/index.js
@@ -1,5 +1,7 @@
 import { toFixedDown, formatBytesToString } from './number';
 import { getChunkCount, getFileNameWithoutExtension } from './upload';
+import { secondsToHms } from './time';
 
 export { toFixedDown, formatBytesToString };
 export { getChunkCount, getFileNameWithoutExtension };
+export { secondsToHms };
